requirements.txt
fastapi
uvicorn
python-dotenv
httpx   
loguru  